Created a header code section to generate code dependent on functions declaration
