+++ harpia/bpGUI/acquisition.py
@@ -313,15 +313,8 @@ class Properties( GladeWindow, S2iCommonProperties):
           dialog.set_current_folder("/home/" + str(os.getenv('USER')) + "/Desktop")
 #Scotti
 
-        filter = gtk.FileFilter()
-        filter.set_name("All Archives")
-        filter.add_pattern("*")
-        dialog.add_filter(filter)
-
-        filter = gtk.FileFilter()
-        filter.set_name("images")
-        filter.add_mime_type("*.jpg")
-        dialog.add_filter(filter)
+        dialog.add_filter(AllFileFilter())
+        dialog.add_filter(JPGFileFilter())
 
         response = dialog.run()
         if response == gtk.RESPONSE_OK:
@@ -350,16 +343,8 @@ class Properties( GladeWindow, S2iCommonProperties):
           dialog.set_current_folder("/home/" + str(os.getenv('USER')) + "/Desktop")
 #Scotti
 
-        filter = gtk.FileFilter()
-        filter.set_name("All Archives")
-        filter.add_pattern("*")
-        dialog.add_filter(filter)
-
-        filter = gtk.FileFilter()
-        filter.set_name("AVI Videos")
-        filter.add_mime_type("*.avi")
-        dialog.add_filter(filter)
-
+        dialog.add_filter(AllFileFilter())
+        dialog.add_filter(AVIFileFilter())
         response = dialog.run()
         if response == gtk.RESPONSE_OK:
             response =  dialog.get_filename()
