+++ harpia/gerador.py
@@ -38,17 +38,17 @@ import gtk
 
 ####################################INCLUDES OF CLASS##########################################
 
-from connection import connection
-from blockTemplate import blockTemplate
-from RunPrg import RunPrg
-
+from harpia.connection import connection
+from harpia.blockTemplate import blockTemplate
+from harpia.RunPrg import RunPrg
 
 ###############################################################################################
 # from popen2 import Popen4
 
-from bpGUI import *
+from harpia.bpGUI import *
 from harpia.utils.XMLUtils import XMLParser
-from constants import *
+from harpia.constants import *
+import s2idirectory
 
 # i18n
 import gettext
@@ -56,8 +56,14 @@ _ = gettext.gettext
 gettext.bindtextdomain(APP, DIR)
 gettext.textdomain(APP)
 
-import s2idirectory
-
+# Global variable to indicate overall behavior of the code generator
+g_bLive = False  # default eh live!!
+g_ShowCount = 0
+g_bFrameRate = 0.1
+usesFindSquares = 0
+g_bCameras = []
+g_bVideo = []  # default eh live!!
+headers = []
 arguments = []
 images = []
 functionCalls = []
@@ -65,6 +71,10 @@ deallocations = []
 outDeallocations = []
 blockList = []
 g_bSaveVideo = []
+usesFindSquares = 0
+usesFindColor = 0
+usesAdjustImage = 0
+usesRads = 0
 
 ErrorLog = 'ErrorLog'
 
@@ -75,9 +85,20 @@ if os.name == "nt":
 else:
     tmpDir = "/tmp/"
 
-
+#----------------------------------------------------------------------
 def __clean_generator():
-    global arguments, images, functionCalls, deallocations, outDeallocations, blockList
+    global headers
+    global arguments
+    global images
+    global functionCalls
+    global deallocations
+    global outDeallocations
+    global blockList
+    global usesFindSquares
+    global usesFindColor
+    global usesAdjustImage
+    global usesRads
+    headers = []
     arguments = []
     images = []
     functionCalls = []
@@ -85,44 +106,34 @@ def __clean_generator():
     outDeallocations = []
     blockList = []
     g_bSaveVideo = []
-
-# Global variable to indicate overall behavior of the code generator
-
-g_bLive = False  # default eh live!!
-
-g_ShowCount = 0
-
-g_bCameras = []
-g_bVideo = []  # default eh live!!
-
-g_bFrameRate = 0.1
-
-usesFindSquares = 0
-usesFindColor = 0
+    usesFindSquares = 0
+    usesFindColor = 0
+    usesAdjustImage = 0
+    usesRads = 0
 
 #----------------------------------------------------------------------
-def inputSizeComply(nInputs, block_number):
-    if nInputs == 2:
-        output_code = 'if(block$bn$_img_i1->width != block$bn$_img_i2->width || ' + \
-                     'block$bn$_img_i1->height != block$bn$_img_i2->height)\n{\n' + \
-                     '	int minW,minH;\n' + \
-                     '	if(block$bn$_img_i1->width > block$bn$_img_i2->width)\n' + \
-                     '		minW = block$bn$_img_i2->width;\n' + \
-                     '	else \n' + \
-                     '		minW = block$bn$_img_i1->width;\n\n' + \
-                     '	if(block$bn$_img_i1->height > block$bn$_img_i2->height)\n' + \
-                     '		minH = block$bn$_img_i2->height;\n' + \
-                     '	else \n' + \
-                     '		minH = block$bn$_img_i1->height;\n\n' + \
-                     '	cvSetImageROI(block$bn$_img_i2, cvRect( 0, 0, minW, minH ));\n' + \
-                     '	cvSetImageROI(block$bn$_img_i1, cvRect( 0, 0, minW, minH ));\n' + \
-                     '	cvSetImageROI(block$bn$_img_o1, cvRect( 0, 0, minW, minH ));\n' + \
-                     '}\n'
-        output_code = output_code.replace("$bn$", str(block_number))
-        return output_code
-    else:
-        return "//Image Sizes match\n"
+def adjust_images_size():
+    return r"""
+// And, Xor, Division, subtraction, sum and multiplication need images with the same size
+void adjust_images_size(IplImage * img1, IplImage * img2, IplImage * img3){
+    if(img1->width != img2->width || img1->height != img2->height){
+	int minW,minH;
+	if(img1->width > img2->width)
+		minW = img2->width;
+	else 
+		minW = img1->width;
 
+	if(img1->height > img2->height)
+		minH = img2->height;
+	else 
+		minH = img1->height;
+
+	cvSetImageROI(img2, cvRect( 0, 0, minW, minH ));
+	cvSetImageROI(img1, cvRect( 0, 0, minW, minH ));
+	cvSetImageROI(img3, cvRect( 0, 0, minW, minH ));
+    }
+}
+"""
 
 #----------------------------------------------------------------------
 def __set_error_log(a_sError):
@@ -133,7 +144,8 @@ def __set_error_log(a_sError):
     Error.write(a_sError)
     Error.close()
 
-def applyWeightsOnConnections(listOfBlocks):
+#----------------------------------------------------------------------
+def __apply_weights_on_connections(listOfBlocks):
     ##For each block on listt:
     returnList = []
     for block in listOfBlocks:
@@ -204,10 +216,7 @@ def parseAndGenerate(dirName, XMLChain, installDirName):
                         tmpConnection.sourceOutput = output.id
                         tmpConnection.destinationInput = output.input
                         tmpConnection.destinationNumber = output.inBlock
-                        # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-                        # !!!!!!cpscotti type-oriented connections...!!!!!
                         tmpConnection.connType = tmpBlock.outTypes[int(tmpConnection.sourceOutput) - 1]
-                        # !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
                         tmpBlock.myConnections.append(tmpConnection)
                     else:
                         tmpConnection.destinationNumber = '--'
@@ -239,13 +248,9 @@ def parseAndGenerate(dirName, XMLChain, installDirName):
             # RollinPathList = []
             tmpList.append(block)
             # RollinPathList.append(block)
-            organizedChain = applyWeightsOnConnections(tmpList)  # ,RollinPathList)
+            organizedChain = __apply_weights_on_connections(tmpList)  # ,RollinPathList)
             while organizedChain != []:
-                organizedChain = applyWeightsOnConnections(organizedChain)  # ,RollinPathList)
-            ###Recursive tests... future functionality
-            # print "Start: "
-            # for block in organizedChain:
-            # print "\t"+str(block.blockNumber)
+                organizedChain = __apply_weights_on_connections(organizedChain)  # ,RollinPathList)
 
     biggestWeight = -1
     for block in blockList:
@@ -256,6 +261,7 @@ def parseAndGenerate(dirName, XMLChain, installDirName):
         activeWeight += 1
         for block in blockList:
             if block.weight == activeWeight:
+                headers.append(block.header)
                 arguments.append(block.functionArguments)
                 images.append(block.imagesIO)
                 functionCalls.append("//Weight: " + str(block.weight) + "\n")
@@ -288,292 +294,11 @@ def parseAndGenerate(dirName, XMLChain, installDirName):
 #include <opencv/highgui.h>
 #include <math.h>
 
-#define PI 3.1415926535898
-double rads(double degs)
-{
-	return (PI/180 * degs);
-}
 		"""
-    global usesFindSquares
-    if usesFindSquares == 1:
-        header += r"""
 
-//Routines to findSquares
-double angle( CvPoint* pt1, CvPoint* pt2, CvPoint* pt0 )
-{
-		double dx1 = pt1->x - pt0->x;
-		double dy1 = pt1->y - pt0->y;
-		double dx2 = pt2->x - pt0->x;
-		double dy2 = pt2->y - pt0->y;
-		return (dx1*dx2 + dy1*dy2)/sqrt((dx1*dx1 + dy1*dy1)*(dx2*dx2 + dy2*dy2) + 1e-10);
-}
+    for header_code in headers:
+        header += header_code
 
-CvSeq* findSquares4( IplImage* img, CvMemStorage* storage, int minArea, int maxArea)
-{
-		CvSeq* contours;
-		int i, c, l, N = 11;
-		int thresh = 50;
-		CvSize sz = cvSize( img->width & -2, img->height & -2 );
-		IplImage* timg = cvCloneImage( img ); // make a copy of input image
-		IplImage* gray = cvCreateImage( sz, 8, 1 ); 
-		IplImage* pyr = cvCreateImage( cvSize(sz.width/2, sz.height/2), 8, 3 );
-		IplImage* tgray;
-		CvSeq* result;
-		double s, t;
-
-		if(minArea == -1)
-			minArea = 0;
-		if(maxArea == -1)
-			maxArea = (img->width * img->height);
-
-		CvSeq* squares = cvCreateSeq( 0, sizeof(CvSeq), sizeof(CvPoint), storage );
-		
-
-		cvSetImageROI( timg, cvRect( 0, 0, sz.width, sz.height ));
-		
-		// down-scale and upscale the image to filter out the noise
-		cvPyrDown( timg, pyr, CV_GAUSSIAN_5x5 );
-		cvPyrUp( pyr, timg, CV_GAUSSIAN_5x5 );
-		tgray = cvCreateImage( sz, 8, 1 );
-		
-		// find squares in every color plane of the image
-		for( c = 0; c < 3; c++ ){
-				// extract the c-th color plane
-				cvSetImageCOI( timg, c+1 );
-				cvCopy( timg, tgray, 0 );
-				for( l = 0; l < N; l++ )
-				{
-						if( l == 0 )
-						{
-								cvCanny( tgray, gray, 0, thresh, 5 );
-								cvDilate( gray, gray, 0, 1 );
-						}
-						else
-						{
-								cvThreshold( tgray, gray, (l+1)*255/N, 255, CV_THRESH_BINARY );
-						}
-						cvFindContours( gray, storage, &contours, sizeof(CvContour),
-								CV_RETR_LIST, CV_CHAIN_APPROX_SIMPLE, cvPoint(0,0) );
-						while( contours )
-						{
-								result = cvApproxPoly( contours, sizeof(CvContour), storage,
-										CV_POLY_APPROX_DP, cvContourPerimeter(contours)*0.02, 0 );
-								if( result->total == 4 &&
-										fabs(cvContourArea(result,CV_WHOLE_SEQ)) > minArea &&
-										fabs(cvContourArea(result,CV_WHOLE_SEQ)) < maxArea &&
-										cvCheckContourConvexity(result) )
-								{
-										s = 0;
-										
-										for( i = 0; i < 5; i++ )
-										{
-												if( i >= 2 )
-												{
-														t = fabs(angle(
-														(CvPoint*)cvGetSeqElem( result, i ),
-														(CvPoint*)cvGetSeqElem( result, i-2 ),
-														(CvPoint*)cvGetSeqElem( result, i-1 )));
-														s = s > t ? s : t;
-												}
-										}
-										if( s < 0.3 )
-												for( i = 0; i < 4; i++ )
-														cvSeqPush( squares,
-																(CvPoint*)cvGetSeqElem( result, i ));
-								}
-								contours = contours->h_next;
-						}
-				}
-		}
-		cvReleaseImage( &gray );
-		cvReleaseImage( &pyr );
-		cvReleaseImage( &tgray );
-		cvReleaseImage( &timg );
-		return squares;
-}
-
-double drawSquares( IplImage* cpy, CvSeq* squares )
-{
-		CvSeqReader reader;
-		int i;
-		cvStartReadSeq( squares, &reader, 0 );
-		for( i = 0; i < squares->total; i += 4 )
-		{
-				CvPoint pt[4], *rect = pt;
-				int count = 4;
-				CV_READ_SEQ_ELEM( pt[0], reader );
-				CV_READ_SEQ_ELEM( pt[1], reader );
-				CV_READ_SEQ_ELEM( pt[2], reader );
-				CV_READ_SEQ_ELEM( pt[3], reader );
-				cvPolyLine( cpy, &rect, &count, 1, 1, CV_RGB(0,255,0), 3, CV_AA, 0 );
-		}
-	return (double)squares->total;
-}
-
-//End of routines to findSquares
-
-			"""
-    global usesFindColor
-    if usesFindColor == 1:
-        header += r"""
-
-int GetColor(IplImage * imagem, int x, int y)
-{
-	return   (int)(((uchar*)(imagem->imageData + imagem->widthStep*y))[x]);
-}
-
-void SetColor(IplImage * imagem, int x, int y, uchar color)
-{
-	((uchar*)(imagem->imageData + imagem->widthStep*y))[x] = color;
-}
-
-
-void CheckImg(IplImage * img, uchar c_value, uchar tolerance)
-{
-	uchar min,max;
-	int y_It,x_It;
-	if((int)c_value < (int)tolerance)
-		tolerance = c_value;
-
-	if(((int)c_value+(int)tolerance) > 255)
-		tolerance = 255 - c_value;
-
-	min = c_value - tolerance;
-	max = c_value + tolerance;
-
-	for(y_It=0;y_It<(img->height);y_It++)
-		for(x_It=0;x_It<(img->width);x_It++)
-		{
-				uchar val;
-				val = GetColor(img,x_It,y_It);
-				if(val >= min && val <= max)
-					SetColor(img,x_It,y_It,255);
-				else
-					SetColor(img,x_It,y_It,0);
-		}
-}
-
-CvPoint GetCenter(IplImage * src, long int * nOfPts)//, long int * numOfPoints)
-{
-	long int numOfMatchingPoints;
-	long int posXsum;
-	long int posYsum;
-	int x_It, y_It;
-	CvPoint Center;
-	
-	posXsum = 0;
-	posYsum = 0;
-	numOfMatchingPoints = 0;
-
-	for(y_It=0;y_It<(src->height);y_It++)
-		for(x_It=0;x_It<(src->width);x_It++)
-			if(GetColor(src,x_It,y_It))
-			{
-				posXsum += x_It;
-				posYsum += y_It;
-				numOfMatchingPoints++;
-			}
-
-	if(numOfMatchingPoints > 0)
-	{
-		Center.x = (int)(posXsum/numOfMatchingPoints);
-		Center.y = (int)(posYsum/numOfMatchingPoints);
-	}
-	else
-		numOfMatchingPoints = -1;
-// 	(*numOfPoints) = numOfMatchingPoints;
-	if(nOfPts)
-		*nOfPts = numOfMatchingPoints;
-
-	return Center;
-}
-
-
-double dist22Points(CvPoint a, CvPoint b)
-{
-	int xD,yD;
-	xD = a.x - b.x;
-	yD = a.y - b.y;
-
-	xD = (xD>0)?xD:-xD;
-	yD = (yD>0)?yD:-yD;
-
-	return (double)(xD*xD + yD*yD);
-}
-
-double GetVariance(IplImage * src,CvPoint center)//, long int * numOfPoints)
-{
-	long int numOfMatchingPoints;
-	double distSquaresSum;
-	double variance;
-	int x_It,y_It;
-
-	numOfMatchingPoints = 0;
-	distSquaresSum = 0.0;
-
-
-	for(y_It=0;y_It<(src->height);y_It++)
-		for(x_It=0;x_It<(src->width);x_It++)
-			if(GetColor(src,x_It,y_It))
-			{
-				numOfMatchingPoints++;
-				distSquaresSum += dist22Points(center,cvPoint(x_It,y_It));
-			}
-
-	if(numOfMatchingPoints)
-		variance = distSquaresSum/numOfMatchingPoints;
-	else
-		variance = -1;
-
-
-	return variance;
-}
-
-long int CheckForColor(IplImage * src, IplImage * dst, uchar * c_value, uchar * tolerance, CvPoint * pointCenter, double * variance)
-{
-	uchar B,B_T,G,G_T,R,R_T;
-	int i;
-	long int numOfPoints;
-	CvPoint centro;
-	IplImage * m_pChans[3] = {NULL,NULL,NULL};
-
-	numOfPoints = 0;
-
-	B = c_value[0];
-	G = c_value[1];
-	R = c_value[2];
-
-	B_T = tolerance[0];
-	G_T = tolerance[1];
-	R_T = tolerance[2];
-
-	for(i=0;i<3;i++)
-		m_pChans[i] = cvCreateImage(cvGetSize(src),IPL_DEPTH_8U, 1);
-
-	cvSplit(src,m_pChans[0],m_pChans[1],m_pChans[2], NULL);
-
-	CheckImg(m_pChans[0],B,B_T);
-	CheckImg(m_pChans[1],G,G_T);
-	CheckImg(m_pChans[2],R,R_T);
-
-	cvAnd(m_pChans[0], m_pChans[1], dst, NULL );
-	cvAnd(m_pChans[2], dst, dst, NULL );
-
-	centro = GetCenter(dst,&numOfPoints);//,&numOfPoints);
-
-	if(numOfPoints != -1)
-		*variance = GetVariance(dst,centro);
-
-	pointCenter->x = centro.x;
-	pointCenter->y = centro.y;
-
-	cvReleaseImage( &m_pChans[0] );
-	cvReleaseImage( &m_pChans[1] );
-	cvReleaseImage( &m_pChans[2] );
-
-	return numOfPoints;
-}
-			"""
     header += "\nint main(int argc, char ** argv)\n{"
 
     declaration = "\n\t//declaration block\n"
@@ -703,20 +428,18 @@ long int CheckForColor(IplImage * src, IplImage * dst, uchar * c_value, uchar * 
 
         o.close()
         i.close()
+
         if g_bLive:
             yield [_("Running, press any key (on the video output window) to terminate."), CompilingErrors]
         else:
             yield [_("Running ..."), CompilingErrors]
 
-        # cpscotti, xunxo bunitinho pra nao travar a interface qndo tive rodando o live =]
         t_oPrg = RunPrg("LD_LIBRARY_PATH=" + installDirName + "/lib/ ./" + codeFilename[:-2])
         t_oPrg.start()
         while t_oPrg.isAlive():
             t_oPrg.join(0.4)
             while gtk.events_pending():
                 gtk.main_iteration(False)
-
-        # fim do "xunx"
 
         ## ERROR LOG
         o = open("RunErrorLog", "r")
